+++ intersection/intersection.h
@@ -11,11 +11,11 @@ typedef struct			s_intersection
 {
 	t_ray				ray;
 	t_vector3			normal;
-    float         diffuse_intensity;
-	float         specular_intensity;
+    double         diffuse_intensity;
+	double         specular_intensity;
 	t_vector3           color;
 	const t_material	*material;
-	float 				shadow_ratio;
+	double 				shadow_ratio;
 	int 				*highlight;
 }						t_intersection;
 
@@ -23,11 +23,11 @@ typedef struct			s_intersection_cl
 {
 	t_ray				ray;
 	t_vector3			normal;
-	float         diffuse_intensity;
-	float         specular_intensity;
+	double         diffuse_intensity;
+	double         specular_intensity;
 	t_vector3           color;
 	t_material			material;
-	float 		shadow_ratio;
+	double 		shadow_ratio;
 	int 				highlight;
 }						t_intersection_cl;
