+++ material/material.c
@@ -1,6 +1,6 @@
 #include "material.h"
 
-t_material			*material_new(t_vector3 color, float diffuse, float specular, float reflect, float refract)
+t_material			*material_new(t_vector3 color, double diffuse, double specular, double reflect, double refract)
 {
 	t_material		*new;
