+++ shape/shape_cylinder.c
@@ -1,10 +1,10 @@
 #include "shape_list.h"
 
 
-static float 		cylinder_caps_intersect(t_shape *shape, t_intersection *intersection)
+static double 		cylinder_caps_intersect(t_shape *shape, t_intersection *intersection)
 {
 	t_cylinder_data	*data;
-	float 			t[2];
+	double 			t[2];
 
 	data = (t_cylinder_data *)shape->data;
 	if (!vector3_dot(&intersection->ray.direction, &data->axis))
@@ -30,11 +30,11 @@ int 			cylinder_intersect
 	(t_shape *shape, t_intersection *intersection)
 {
 	t_cylinder_data	*data;
-	float 			k[3];
-	float 			discriminant;
-	float 			t[2];
+	double 			k[3];
+	double 			discriminant;
+	double 			t[2];
 	t_vector3		temp[2];
-	float 			angle[2];
+	double 			angle[2];
 
 	data = (t_cylinder_data *)shape->data;
 	temp[0] = vector3_sub(&intersection->ray.origin, &data->bottom);
@@ -76,7 +76,7 @@ static void			cylinder_move(t_shape *shape, t_vector3 move)
 }
 
 t_shape				*shape_cylinder
-	(t_vector3 top, t_vector3 bottom, float radius, const t_material *material)
+	(t_vector3 top, t_vector3 bottom, double radius, const t_material *material)
 {
 	t_shape			*shape;
 	t_cylinder_data	*data;
